+++ dropwizard-auth/src/test/java/io/dropwizard/auth/basic/BasicAuthProviderTest.java
@@ -1,14 +1,14 @@
 package io.dropwizard.auth.basic;
 
 import com.codahale.metrics.MetricRegistry;
-import com.google.common.base.Optional;
-import io.dropwizard.auth.Auth;
-import io.dropwizard.auth.AuthFactory;
+import io.dropwizard.auth.AuthDynamicFeature;
+import io.dropwizard.auth.AuthFilter;
 import io.dropwizard.auth.AuthResource;
-import io.dropwizard.auth.AuthenticationException;
 import io.dropwizard.auth.Authenticator;
+import io.dropwizard.auth.util.AuthUtil;
 import io.dropwizard.jersey.DropwizardResourceConfig;
 import io.dropwizard.logging.LoggingFactory;
+import org.glassfish.jersey.server.filter.RolesAllowedDynamicFeature;
 import org.glassfish.jersey.servlet.ServletProperties;
 import org.glassfish.jersey.test.DeploymentContext;
 import org.glassfish.jersey.test.JerseyTest;
@@ -17,19 +17,23 @@ import org.glassfish.jersey.test.grizzly.GrizzlyWebTestContainerFactory;
 import org.glassfish.jersey.test.spi.TestContainerException;
 import org.glassfish.jersey.test.spi.TestContainerFactory;
 import org.junit.Test;
-
+import javax.annotation.security.RolesAllowed;
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 import javax.ws.rs.WebApplicationException;
+import javax.ws.rs.container.ContainerRequestFilter;
+import javax.ws.rs.core.Context;
 import javax.ws.rs.core.HttpHeaders;
 import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.SecurityContext;
+import java.security.Principal;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.failBecauseExceptionWasNotThrown;
 
-
 public class BasicAuthProviderTest extends JerseyTest {
+    final private static String VALID_ROLE = "ADMIN";
     static {
         LoggingFactory.bootstrap();
     }
@@ -56,6 +60,40 @@ public class BasicAuthProviderTest extends JerseyTest {
             assertThat(e.getResponse().getStatus()).isEqualTo(401);
             assertThat(e.getResponse().getHeaders().get(HttpHeaders.WWW_AUTHENTICATE))
                     .containsOnly("Basic realm=\"realm\"");
+        }
+    }
+
+    @Test
+    public void resourceWithoutAuth200() {
+        assertThat(target("/test/noauth").request()
+                .get(String.class))
+                .isEqualTo("hello");
+    }
+
+    @Test
+    public void resourceWithAuthNotRequired200() {
+        assertThat(target("/test/authnotrequired").request()
+                .get(String.class))
+                .isEqualTo("No Principal");
+    }
+
+    @Test
+    public void resourceWithDenyAllAndNoAuth401() {
+        try {
+            target("/test/denied").request().get(String.class);
+        } catch (WebApplicationException e) {
+            assertThat(e.getResponse().getStatus()).isEqualTo(401);
+        }
+    }
+
+    @Test
+    public void resourceWithDenyAllAndAuth403() {
+        try {
+            target("/test/denied").request()
+                    .header(HttpHeaders.AUTHORIZATION, "Basic Z29vZC1ndXk6c2VjcmV0")
+                    .get(String.class);
+        } catch (WebApplicationException e) {
+            assertThat(e.getResponse().getStatus()).isEqualTo(403);
         }
     }
 
@@ -97,9 +135,10 @@ public class BasicAuthProviderTest extends JerseyTest {
     @Path("/test/")
     @Produces(MediaType.TEXT_PLAIN)
     public static class ExampleResource {
+        @RolesAllowed({VALID_ROLE})
         @GET
-        public String show(@Auth String principal) {
-            return principal;
+        public String show(@Context SecurityContext context) {
+            return context.getUserPrincipal().getName();
         }
     }
 
@@ -107,21 +146,19 @@ public class BasicAuthProviderTest extends JerseyTest {
         public BasicAuthTestResourceConfig() {
             super(true, new MetricRegistry());
 
-            final Authenticator<BasicCredentials, String> authenticator = new Authenticator<BasicCredentials, String>() {
-                @Override
-                public Optional<String> authenticate(BasicCredentials credentials) throws AuthenticationException {
-                    if ("good-guy".equals(credentials.getUsername()) &&
-                            "secret".equals(credentials.getPassword())) {
-                        return Optional.of("good-guy");
-                    }
-                    if ("bad-guy".equals(credentials.getUsername())) {
-                        throw new AuthenticationException("CRAP");
-                    }
-                    return Optional.absent();
-                }
-            };
-            register(AuthFactory.binder(new BasicAuthFactory<>(authenticator, "realm", String.class)));
+            register(new AuthDynamicFeature(getAuthFilter()));
+            register(RolesAllowedDynamicFeature.class);
             register(AuthResource.class);
+        }
+
+        private ContainerRequestFilter getAuthFilter() {
+            final String validUser = "good-guy";
+
+            BasicCredentialAuthFilter.Builder<Principal, Authenticator<BasicCredentials, Principal>> builder
+                    = new BasicCredentialAuthFilter.Builder<>();
+            builder.setSecurityContextFunction(AuthUtil.<AuthFilter.Tuple, SecurityContext>getSecurityContextProviderFunction(validUser, VALID_ROLE));
+            builder.setAuthenticator(AuthUtil.<BasicCredentials, Principal>getTestAuthenticatorBasicCredential(validUser));
+            return builder.buildAuthHandler();
         }
     }
 }