+++ dropwizard-example/src/main/java/com/example/helloworld/core/User.java
@@ -1,6 +1,8 @@
 package com.example.helloworld.core;
 
-public class User {
+import java.security.Principal;
+
+public class User implements Principal {
     private final String name;
 
     public User(String name) {
@@ -9,5 +11,9 @@ public class User {
 
     public String getName() {
         return name;
+    }
+
+    public int getId() {
+        return (int) (Math.random() * 100);
     }
 }